improve runtime
