+++ crates/runtime/src/lib.rs
@@ -1,11 +1,13 @@
 //! # WebAssembly Runtime
 
+mod cli;
+mod compiler;
 mod runtime;
 mod service;
 mod trace;
-mod cli;
 
+pub use self::cli::*;
+pub use self::compiler::*;
 pub use self::runtime::*;
 pub use self::service::*;
 pub use self::trace::*;
-pub use self::cli::*;
