+++ src/bin/wasmgrid.rs
@@ -1,22 +1,9 @@
-#![feature(impl_trait_in_assoc_type)]
-
 //! # Wasmgrid CLI
 
 use dotenv::dotenv;
-// #[cfg(feature = "messaging")]
-// use wasmgrid::messaging;
-// #[cfg(feature = "rpc")]
-// use wasmgrid::rpc;
-// #[cfg(feature = "vault")]
-// use crate::vault;
 use runtime::{Cli, Command, Parser};
 use tracing_subscriber::{EnvFilter, FmtSubscriber};
-#[cfg(feature = "http")]
-use wasmgrid::http;
-// #[cfg(feature = "jsondb")]
-// use wasmgrid::jsondb;
-#[cfg(feature = "keyvalue")]
-use wasmgrid::keyvalue;
+use wasmgrid::*;
 
 #[tokio::main]
 pub async fn main() -> wasmtime::Result<()> {
@@ -26,8 +13,6 @@ pub async fn main() -> wasmtime::Result<()> {
     let subscriber =
         FmtSubscriber::builder().with_env_filter(EnvFilter::from_default_env()).finish();
     tracing::subscriber::set_global_default(subscriber)?;
-
-    // runtime::init()?;
 
     match Cli::parse().command {
         Command::Compile { wasm, output } => {
@@ -51,4 +36,3 @@ pub async fn main() -> wasmtime::Result<()> {
         }
     }
 }
-
